Woot, task for synthesizing NO2 and O3 data created
    run_trends.scd modified to load play_funcs
    trend_data_funcs.scd modified so that the global stats include a field
describing the minimum number of data points in time for each variable, to
be used in a loop to reset the counter.

NEW FILES:
    play_funcs.scd - contains all the functions used to initialize the
synth driver task. The task itself should be given the ~task_driver_func
as the function for it to use. All the inputs to that function must be
objects that get passed by reference so that they can be modified later.

    play_main.scd - calls the initialization functions in play_funcs.
Stores them in global variables, but the global variable names are not
references anywhere else yet, so they can be changed. Control of options
for the task should be done through the ~taskOpts dictionary; note that
if you want to add options there, you'll need to add them to the "req_opts"
variable in ~task_driver_func, as it checks that exactly those options are
keys in the dictionary.