Buttons to turn on or off all sites in a category implemented.
Needed to move the controls view down and expand the whole window.
Also, the ~base_path should be independent of the computer it's on now.
By reading the path that the run_trends file is in and making the base
path based on that, it should catch the different file structure for our computers.