Time stepping controls added
Ability to step by month or year (or rewind to beginning) added.
Have not tested yet what happens if they are pressed while automatically
advancing.
I got this to work by making the date index global, incrementing or
decrementing it with the buttons (ensuring it stays within the proper
range) then restarting the main task to make sure that the entire loop
executes. A new option, "oneStep" for the taskOpts dictionary controls
the behavior at the end of the loop, telling the main task to stop
if "oneStep" is true.

Also I flip-flopped the loop/no loop button states b/c true == 1 and
false == 0, so now updating the GUI based on the value in the taskOpts
dict does not cause loop to flip on and off every time a step is taken.