Some typos fixed
Flipped the order of paragraphs in sect. 2.3 a bit; thought maybe
it would make more sense to talk about the different sounds after
we'd introduced the panning approaches so that it makes sense that
the ability to localize the sound's location in the panning field
is important. If you don't agree, feel free to switch it back ;)